requirements.txt
fastapi
uvicorn[standard]
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pypdf2
python-multipart
python-dotenv
requests
pytest
scikit-learn
loguru
pinecone-client
tqdm
